a lot of changes towards 2.10.0
